requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
asyncpg==0.29.0
akshare==1.17.42
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
redis==5.0.1
pandas==2.1.4
numpy==1.26.2
aiofiles==23.2.1